feat: 테스트 편의성을 위해 LikeService를 인터페이스로 변경
